refactor(wasm): merge network/message
